Refactored UI elements representing ruleset XML dom elements to help make things be more generic and speed up the dev process of building the UI elements.
